Retornar se o v;say não tiver args

+++ commands/SayCommand.js
@@ -2,6 +2,7 @@ const Discord = require("discord.js");
 
 exports.run = async (vary, message, args) => {
 if (!message.member.hasPermission("MANAGE_MESSAGES")) return message.reply("Permissão `Gerenciar mensagens` não encontrada em seu cargo!")
+if (args[0] == null) {return message.channel.send(`Digite algo para eu mandar!`)}
 const sayeMessage = args.join(" ");
     message.delete().catch(O_o=>{});
     message.channel.send(sayeMessage);
